+++ RayTracer.MathLibrary.Tests/MathHelperTests.cs
@@ -12,30 +12,30 @@ namespace RayTracer.MathLibrary.Tests
 
         [Test]
         [TestCase(0f, 0f, 0f)]
-        public void CreateObjectIn3DSpace_WhenCalledWithPointCordinates_DoesCreatePoint(float x, float y, float z)
+        public void CreateObjectIn3DSpace_WhenCalledWithPointCordinates_DoesCreatePoint(double x, double y, double z)
         {
-            Assert.IsInstanceOf<Point3D>(MathHelper.CreateObjectIn3DSpace(Tuple.Create<float, float, float, float>(x, y, z, 1.0f)));
+            Assert.IsInstanceOf<Point3D>(MathHelper.CreateObjectIn3DSpace(Tuple.Create<double, double, double, double>(x, y, z, 1.0f)));
         }
 
         [Test]
         [TestCase(0f, 0f, 0f)]
-        public void CreateObjectIn3DSpace_WhenCalledWithPointCordinates_DoesNotCreateVector(float x, float y, float z)
+        public void CreateObjectIn3DSpace_WhenCalledWithPointCordinates_DoesNotCreateVector(double x, double y, double z)
         {
-            Assert.IsNotInstanceOf<Vector3>(MathHelper.CreateObjectIn3DSpace(Tuple.Create<float, float, float, float>(x, y, z, 1.0f)));
+            Assert.IsNotInstanceOf<Vector3>(MathHelper.CreateObjectIn3DSpace(Tuple.Create<double, double, double, double>(x, y, z, 1.0f)));
         }
 
         [Test]
         [TestCase(0f, 0f, 0f)]
-        public void CreateObjectIn3DSpace_WhenCalledWithVectorCordinates_DoesCreateVector(float x, float y, float z)
+        public void CreateObjectIn3DSpace_WhenCalledWithVectorCordinates_DoesCreateVector(double x, double y, double z)
         {
-            Assert.IsInstanceOf<Vector3>(MathHelper.CreateObjectIn3DSpace(Tuple.Create<float, float, float, float>(x, y, z, 0.0f)));
+            Assert.IsInstanceOf<Vector3>(MathHelper.CreateObjectIn3DSpace(Tuple.Create<double, double, double, double>(x, y, z, 0.0f)));
         }
 
         [Test]
         [TestCase(0f, 0f, 0f)]
-        public void CreateObjectIn3DSpace_WhenCalledWithVectorCordinates_DoesNotCreatePoint(float x, float y, float z)
+        public void CreateObjectIn3DSpace_WhenCalledWithVectorCordinates_DoesNotCreatePoint(double x, double y, double z)
         {
-            Assert.IsNotInstanceOf<Point3D>(MathHelper.CreateObjectIn3DSpace(Tuple.Create<float, float, float, float>(x, y, z, 0.0f)));
+            Assert.IsNotInstanceOf<Point3D>(MathHelper.CreateObjectIn3DSpace(Tuple.Create<double, double, double, double>(x, y, z, 0.0f)));
         }
 
         [Test]
@@ -45,7 +45,7 @@ namespace RayTracer.MathLibrary.Tests
         [TestCase(true, 0.000009f, 0f, 0.00001f)]
         [TestCase(false, 0f, 0.11f, 0.1f)]
         [TestCase(false, 1f, 0.09f, 0.1f)]
-        public void AreNumbersEqual_WhenCalled_ReturnIfNumbersAreEqual(bool expected, float first, float second, float epsilon)
+        public void AreNumbersEqual_WhenCalled_ReturnIfNumbersAreEqual(bool expected, double first, double second, double epsilon)
         {
             Assert.AreEqual(expected, MathHelper.AreNumbersEqual(first, second, epsilon));
         }
