+++ RayTracer.MathLibrary/IObjectIn3DSpace.cs
@@ -6,10 +6,10 @@ namespace RayTracer.MathLibrary
 {
     public interface IObjectIn3DSpace
     {
-        public float X { get; set; }
+        public double X { get; set; }
 
-        public float Y { get; set; }
+        public double Y { get; set; }
 
-        public float Z { get; set; }
+        public double Z { get; set; }
     }
 }
